# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API",
        "description": "Products API for CRUD operations ",
        "contact": {
            "email": "innotuyi81@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "a3cd7435fd5a6e8c78703f8db1faa624",
                "responses": {
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success|Returns Products list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Iphone 14"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "this is products"
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": "6"
                                            },
                                            "price": {
                                                "type": "decimal",
                                                "example": "65.5"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "a696b4db38af5dea1a62113749aa0524",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "decimal"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Bebe",
                                    "description": "this is description",
                                    "price": "101",
                                    "quantity": "4"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Bebe"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "this is description"
                                        },
                                        "price": {
                                            "type": "string",
                                            "example": "40"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": "8"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-02-23T07:07:54.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-02-23T07:07:54.000000Z"
                                        },
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/product/search/{name}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search for Related name in DataBase",
                "operationId": "c3659d195499799b140702dca1a1fc76",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "laptop"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "this is portable computer"
                                        },
                                        "price": {
                                            "type": "string",
                                            "example": "808"
                                        },
                                        "quantity": {
                                            "type": "string",
                                            "example": "38"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "UnAuthanticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Returns when there is server problem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}